+++ tests/Mscc.GenerativeAI/VertexAi_Gemini15Pro_Should.cs
@@ -245,16 +245,10 @@ namespace Test.Mscc.GenerativeAI
             // Arrange
             var prompt = @"User input: I like bagels.
 Answer:";
-            var parts = new List<IPart>
+            var systemInstruction = new Content
             {
-                new TextData { Text = "You are a helpful language translator." },
-                new TextData { Text = "Your mission is to translate text in English to French." }
+                Parts = new() { new TextData { Text = "You are a friendly pirate. Speak like one. Your mission is to translate text in English to French." }}
             };
-            var systemInstruction = new List<Content> { new Content
-            {
-                Role = Role.System,
-                Parts = parts
-            }};
             var generationConfig = new GenerationConfig() 
                 { 
                     Temperature = 0.9f,
@@ -282,7 +276,7 @@ Answer:";
             response.Should().NotBeNull();
             response.Candidates.Should().NotBeNull().And.HaveCount(1);
             response.Text.Should().NotBeNull();
-            _output.WriteLine($"Answer:\n{response?.Text}");
+            _output.WriteLine($"{prompt} {response?.Text}");
             _output.WriteLine($"Usage metadata: {response.UsageMetadata.TotalTokenCount}");
             _output.WriteLine($"Finish reason: {response.Candidates[0].FinishReason}");
             _output.WriteLine($"Safety settings: {response.Candidates[0].SafetyRatings}");